+++ src/Model/Feed/FeedConfiguration.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Model\Feed;
+
+final class FeedConfiguration
+{
+    public readonly string $type;
+    public readonly string $url;
+    public readonly string $timezone;
+    public readonly string $rootPointer;
+    public readonly string $dateFormat;
+    /** @var array list<string> */
+    public readonly array $mapping;
+    /** @var array list<string> */
+    public readonly array $defaults;
+
+    public function __toString(): string
+    {
+        $output = [];
+        $output[] = 'Type: '.$this->type;
+        $output[] = 'Url: '.$this->url;
+
+        return implode("\n", $output);
+    }
+}