+++ src/Services/Feeds/Mapper/FeedMapperInterface.php
@@ -2,23 +2,21 @@
 
 namespace App\Services\Feeds\Mapper;
 
+use App\Model\Feed\FeedConfiguration;
 use App\Model\Feed\FeedItem;
+use CuyZ\Valinor\Mapper\MappingError;
 
 interface FeedMapperInterface
 {
     /**
-     * Map data array into type FeedItem object.
-     *
-     * @todo: Make mapping dynamic with feed configuration.
-     *
-     * @todo: Convert mapping array to value object.
+     * Transform raw feed item array into typed FeedItem object.
      *
      * @param array  $data
-     *  Array with data from feed
-     * @param array  $mapping
-     *   Field mapping from data to FeedItem
-     * @param string $dateFormat
-     *   Date format for date fields
+     *   Array with data from feed
+     * @param FeedConfiguration $configuration
+     *   Feed configuration.
+     *
+     * @throws MappingError
      */
-    public function getFeedItemFromArray(array $data, array $mapping, string $dateFormat = 'Y-m-d\TH:i:s'): FeedItem;
+    public function getFeedItemFromArray(array $data, FeedConfiguration $configuration): FeedItem;
 }